Move debug functionality out of prototype state and into debug menu
