+++ src/Core/Player/HUD.cpp
@@ -14,10 +14,20 @@ void HUD::update(float dt) { gui->update(dt); }
 void HUD::addToOverlay() {
     gui = bl::gui::GUI::create(owner.getCurrentWorld(), owner, bl::gui::LinePacker::create());
     debugMenu.init(*gui);
-    gui->addToOverlay();
+    gui->addToOverlay(nullptr, false);
+    gui->setRegion({0.f, 0.f, 0.f, 0.f});
+    bl::event::Dispatcher::subscribe(this);
 }
 
-void HUD::removeFromOverlay() { gui.reset(); }
+void HUD::removeFromOverlay() {
+    gui.reset();
+    bl::event::Dispatcher::unsubscribe(this);
+}
+
+void HUD::observe(const sf::Event& event) {
+    if (gui && gui->processEvent(event)) { return; }
+    if (debugMenu.processEvent(event)) { return; }
+}
 
 } // namespace player
 } // namespace core
